added alpha to rgb rain

+++ src/main/kotlin/frc/team6502/kyberlib/lighting/animations/AnimationBlink.kt
@@ -5,10 +5,10 @@ import kotlin.math.PI
 import kotlin.math.roundToInt
 import kotlin.math.sin
 
-class AnimationBlink(val color: Color, val ticksPerCycle: Double) : LEDAnimation() {
+class AnimationBlink(val color: Color, val ticksPerCycle: Int) : LEDAnimation() {
 
     override fun getBuffer(ticks: Int, length: Int): List<Color> {
-        val brightness = ((sin(2 * PI * (ticks % ticksPerCycle) / ticksPerCycle) + 1) / 2).roundToInt()
+        val brightness = ((sin(2 * PI * (ticks % ticksPerCycle) / ticksPerCycle.toDouble()) + 1) / 2).roundToInt()
         return Array(length) {
             Color(color.red / 255F, color.green / 255F, color.blue / 255F, brightness.toFloat())
         }.toMutableList()
